+++ addOns/openapi/src/main/java/org/zaproxy/zap/extension/openapi/generators/BodyGenerator.java
@@ -19,6 +19,8 @@
  */
 package org.zaproxy.zap.extension.openapi.generators;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import io.swagger.v3.core.util.Json;
 import io.swagger.v3.oas.models.media.ArraySchema;
 import io.swagger.v3.oas.models.media.ComposedSchema;
 import io.swagger.v3.oas.models.media.Schema;
@@ -100,9 +102,21 @@ public class BodyGenerator {
     }
 
     private String generateFromArraySchema(ArraySchema schema) {
-        StringBuilder json = new StringBuilder();
-        json.append(generate(schema.getItems()));
-        return createJsonArrayWith(json.toString());
+        if (schema.getExample() instanceof String) {
+            return (String) schema.getExample();
+        }
+
+        if (schema.getExample() instanceof Iterable) {
+            try {
+                return Json.mapper().writeValueAsString(schema.getExample());
+            } catch (JsonProcessingException e) {
+                LOG.warn(
+                        "Failed to encode Example Object. Falling back to default example generation",
+                        e);
+            }
+        }
+
+        return createJsonArrayWith(generate(schema.getItems()));
     }
 
     @SuppressWarnings("rawtypes")