+++ addOns/openapi/src/test/java/org/zaproxy/zap/extension/openapi/v3/OpenApiUnitTest.java
@@ -154,6 +154,53 @@ public class OpenApiUnitTest extends AbstractServerTest {
     }
 
     @Test
+    public void shouldExplorePetStoreYamlWithExamples()
+            throws NullPointerException, IOException, SwaggerException {
+        String test = "/PetStoreYamlExamples/";
+        String defnName = "defn.yaml";
+
+        this.nano.addHandler(
+                new DefnServerHandler(test, defnName, "PetStore_defn_with_examples.yaml"));
+
+        Requestor requestor = new Requestor(HttpSender.MANUAL_REQUEST_INITIATOR);
+        HttpMessage defnMsg = this.getHttpMessage(test + defnName);
+        SwaggerConverter converter =
+                new SwaggerConverter(
+                        requestor.getResponseBody(defnMsg.getRequestHeader().getURI()), null);
+        // No parsing errors
+        assertThat(converter.getErrorMessages(), is(empty()));
+
+        final Map<String, String> accessedUrls = new HashMap<String, String>();
+        RequesterListener listener =
+                new RequesterListener() {
+                    @Override
+                    public void handleMessage(HttpMessage message, int initiator) {
+                        accessedUrls.put(
+                                message.getRequestHeader().getMethod()
+                                        + " "
+                                        + message.getRequestHeader().getURI().toString(),
+                                message.getRequestBody().toString());
+                    }
+                };
+        requestor.addListener(listener);
+        requestor.run(converter.getRequestModels());
+
+        assertTrue(
+                "Should use OpenAPI Example Values in URL Path when crawling urls",
+                accessedUrls.containsKey(
+                        "GET http://localhost:"
+                                + this.nano.getListeningPort()
+                                + "/PetStore/store/order/42424242"));
+
+        assertTrue(
+                "Should use OpenAPI Example Values in URL Query when crawling urls",
+                accessedUrls.containsKey(
+                        "GET http://localhost:"
+                                + this.nano.getListeningPort()
+                                + "/PetStore/user/login?username=kermit&password=thefrog"));
+    }
+
+    @Test
     public void shouldExplorePetStoreWithDefaultUrl()
             throws NullPointerException, IOException, SwaggerException {
         // Given